+++ imod_coupler/drivers/dfm_metamod/dfm_metamod.py
@@ -1,329 +1,319 @@
+""" MetaMod: the coupling between MetaSWAP and MODFLOW 6
+
+description:
+
+"""
+from __future__ import annotations
+
 from pathlib import Path
-from typing import Optional, Union
+from typing import Any, Dict
 
-import tomli_w
+import numpy as np
+from loguru import logger
+from numpy.typing import NDArray
+from scipy.sparse import csr_matrix, dia_matrix
+from xmipy import XmiWrapper
+from bmi.wrapper import BMIWrapper
 
-from imod.couplers.metamod.node_svat_mapping import NodeSvatMapping
-from imod.couplers.metamod.rch_svat_mapping import RechargeSvatMapping
-from imod.couplers.metamod.wel_svat_mapping import WellSvatMapping
-from imod.mf6 import Modflow6Simulation
-from imod.msw import GridData, MetaSwapModel, Sprinkling
-from hydrolib.core.io.mdu.models import FMModel
+from imod_coupler.config import BaseConfig
+from imod_coupler.drivers.dfm_metamod.config import Coupling, DfmMetaModConfig
+from imod_coupler.drivers.driver import Driver
+from imod_coupler.utils import create_mapping
+import os
 
-class Dfm_MetaMod:
-    """
-    The Dfm_MetaMod class creates the necessary input files for coupling MetaSWAP,
-    MODFLOW 6 and Dflow-FM.
 
-    Parameters
-    ----------
-    msw_model : MetaSwapModel
-        The MetaSWAP model that should be coupled.
-    mf6_simulation : Modflow6Simulation
-        The Modflow6 simulation that should be coupled.
-    dfm_model: 
-        DFlow-FM model that should be coupled
-    mf6_rch_pkgkey: str
-        Key of Modflow 6 recharge package to which MetaSWAP is coupled.
-    mf6_wel_pkgkey: str or None
-        Optional key of Modflow 6 well package to which MetaSWAP sprinkling is
-        coupled.
-    """
+class DfmMetaMod(Driver):
+    """The driver coupling DFLOW-FM, MetaSWAP and MODFLOW 6"""
 
-    _toml_name = "imod_coupler.toml"
-    _modflow6_model_dir = "Modflow6"
-    _metaswap_model_dir = "MetaSWAP"
-    _dfm_model_dir = "dflow-fm"
+    name: str = "dfm_metamod"  # name of the driver
+    base_config: BaseConfig  # the parsed information from the configuration file
+    dfm_metamod_config: DfmMetaModConfig  # the parsed information from the configuration file specific to MetaMod
+    coupling: Coupling  # the coupling information
+
+    timing: bool  # true, when timing is enabled
+    mf6: XmiWrapper  # the MODFLOW 6 XMI kernel
+    msw: XmiWrapper  # the MetaSWAP XMI kernel
+    dfm: BMIWrapper  # the dflow-fm BMI kernel
+
+    max_iter: NDArray[Any]  # max. nr outer iterations in MODFLOW kernel
+    delt: float  # time step from MODFLOW 6 (leading)
+
+    mf6_head: NDArray[Any]  # the hydraulic head array in the coupled model
+    mf6_recharge: NDArray[Any]  # the coupled recharge array from the RCH package
+    mf6_storage: NDArray[Any]  # the specific storage array (ss)
+    mf6_has_sc1: bool  # when true, specific storage in mf6 is given as a storage coefficient (sc1)
+    mf6_area: NDArray[Any]  # cell area (size:nodes)
+    mf6_top: NDArray[Any]  # top of cell (size:nodes)
+    mf6_bot: NDArray[Any]  # bottom of cell (size:nodes)
+
+    mf6_sprinkling_wells: NDArray[Any]  # the well data for coupled extractions
+    msw_head: NDArray[Any]  # internal MetaSWAP groundwater head
+    msw_volume: NDArray[Any]  # unsaturated zone flux (as a volume!)
+    msw_storage: NDArray[Any]  # MetaSWAP storage coefficients (MODFLOW's sc1)
+    msw_time: float  # MetaSWAP current time
+
+    dflowfm_stage:  NDArray[Any] 
+
+    # dictionary with mapping tables for mod=>msw coupling
+    map_mod2msw: Dict[str, csr_matrix] = {}
+    # dictionary with mapping tables for msw=>mod coupling
+    map_msw2mod: Dict[str, csr_matrix] = {}
+    # dict. with mask arrays for mod=>msw coupling
+    mask_mod2msw: Dict[str, NDArray[Any]] = {}
+    # dict. with mask arrays for msw=>mod coupling
+    mask_msw2mod: Dict[str, NDArray[Any]] = {}
 
     def __init__(
-        self,
-        msw_model: MetaSwapModel,
-        mf6_simulation: Modflow6Simulation,
-        dfm_model: FMModel,
-        mf6_rch_pkgkey: str,
-        mf6_wel_pkgkey: Optional[str] = None,
+        self, base_config: BaseConfig, config_dir: Path, driver_dict: Dict[str, Any]
     ):
-        self.msw_model = msw_model
-        self.mf6_simulation = mf6_simulation
-        self.mf6_rch_pkgkey = mf6_rch_pkgkey
-        self.mf6_wel_pkgkey = mf6_wel_pkgkey
-        self.dfm_model = dfm_model
-        self.is_sprinkling = self._check_coupler_and_sprinkling()
+        """Constructs the `DfmMetaMod` object"""
+        self.base_config = base_config
+        self.dfm_metamod_config = DfmMetaModConfig(config_dir, **driver_dict)
+        self.coupling = self.dfm_metamod_config.coupling[
+            0
+        ]  # Adapt as soon as we have multimodel support
 
-    def _check_coupler_and_sprinkling(self):
-        mf6_rch_pkgkey = self.mf6_rch_pkgkey
-        mf6_wel_pkgkey = self.mf6_wel_pkgkey
+    def initialize(self) -> None:
+        self.mf6 = XmiWrapper(
+            lib_path=self.dfm_metamod_config.kernels.modflow6.dll,
+            lib_dependency=self.dfm_metamod_config.kernels.modflow6.dll_dep_dir,
+            working_directory=self.dfm_metamod_config.kernels.modflow6.work_dir,
+            timing=self.base_config.timing,
+        )
+        self.msw = XmiWrapper(
+            lib_path=self.dfm_metamod_config.kernels.metaswap.dll,
+            lib_dependency=self.dfm_metamod_config.kernels.metaswap.dll_dep_dir,
+            working_directory=self.dfm_metamod_config.kernels.metaswap.work_dir,
+            timing=self.base_config.timing,
+        )
+        
+        #================
+        # modifying the path here should not be necessary
+        os.environ["PATH"] = ( os.path.dirname(self.dfm_metamod_config.kernels.dflowfm.dll) + os.pathsep + os.environ["PATH"] )
+        #================
+        mdu_name = self.coupling.dict()["dfm_model"]
+        dflowfm_input = self.dfm_metamod_config.kernels.dflowfm.work_dir.joinpath(mdu_name)
+        self.dfm = BMIWrapper( engine="dflowfm", configfile=dflowfm_input)
+        
+        # Print output to stdout
+        self.mf6.set_int("ISTDOUTTOFILE", 0)
+        self.mf6.initialize()
+        self.msw.initialize()
+        self.dfm.initialize()
+        self.log_version()
+        self.couple()
 
-        gwf_names = self._get_gwf_modelnames()
+    def log_version(self) -> None:
+        logger.info(f"MODFLOW version: {self.mf6.get_version()}")
+        logger.info(f"MetaSWAP version: {self.msw.get_version()}")
+        logger.info(f"Dflow FM version: version fetching not implemented in BMI")
 
-        # Assume only one groundwater flow model
-        # FUTURE: Support multiple groundwater flow models.
-        gwf_model = self.mf6_simulation[gwf_names[0]]
+    def couple(self) -> None:
+        """Couple Modflow and Metaswap"""
+        # get some 'pointers' to MF6 and MSW internal data
+        mf6_head_tag = self.mf6.get_var_address("X", self.coupling.mf6_model)
+        mf6_recharge_tag = self.mf6.get_var_address(
+            "BOUND", self.coupling.mf6_model, self.coupling.mf6_msw_recharge_pkg
+        )
+        mf6_storage_tag = self.mf6.get_var_address("SS", self.coupling.mf6_model, "STO")
+        mf6_is_sc1_tag = self.mf6.get_var_address(
+            "ISTOR_COEF", self.coupling.mf6_model, "STO"
+        )
+        mf6_area_tag = self.mf6.get_var_address("AREA", self.coupling.mf6_model, "DIS")
+        mf6_top_tag = self.mf6.get_var_address("TOP", self.coupling.mf6_model, "DIS")
+        mf6_bot_tag = self.mf6.get_var_address("BOT", self.coupling.mf6_model, "DIS")
+        mf6_max_iter_tag = self.mf6.get_var_address("MXITER", "SLN_1")
 
-        if mf6_rch_pkgkey not in gwf_model.keys():
-            raise ValueError(
-                f"No package named {mf6_rch_pkgkey} detected in Modflow 6 model. "
-                "iMOD_coupler requires a Recharge package."
+        self.mf6_head = self.mf6.get_value_ptr(mf6_head_tag)
+        # NB: recharge is set to first column in BOUND
+        self.mf6_recharge = self.mf6.get_value_ptr(mf6_recharge_tag)[:, 0]
+        self.mf6_storage = self.mf6.get_value_ptr(mf6_storage_tag)
+        self.mf6_has_sc1 = self.mf6.get_value_ptr(mf6_is_sc1_tag)[0] != 0
+        self.mf6_area = self.mf6.get_value_ptr(mf6_area_tag)
+        self.mf6_top = self.mf6.get_value_ptr(mf6_top_tag)
+        self.mf6_bot = self.mf6.get_value_ptr(mf6_bot_tag)
+        self.max_iter = self.mf6.get_value_ptr(mf6_max_iter_tag)[0]
+
+        self.msw_head = self.msw.get_value_ptr("dhgwmod")
+        self.msw_volume = self.msw.get_value_ptr("dvsim")
+        self.msw_storage = self.msw.get_value_ptr("dsc1sim")
+
+        # create a lookup, with the svat tuples (id, lay) as keys and the
+        # metaswap internal indexes as values
+        svat_lookup = {}
+        msw_mod2svat_file = self.msw.working_directory / "mod2svat.inp"
+        if msw_mod2svat_file.is_file():
+            svat_data: NDArray[np.int32] = np.loadtxt(
+                msw_mod2svat_file, dtype=np.int32, ndmin=2
             )
-
-        sprinkling_key = self.msw_model._get_pkg_key(Sprinkling, optional_package=True)
-
-        sprinkling_in_msw = sprinkling_key is not None
-        sprinkling_in_mf6 = mf6_wel_pkgkey in gwf_model.keys()
-
-        if sprinkling_in_msw and not sprinkling_in_mf6:
-            raise ValueError(
-                f"No package named {mf6_wel_pkgkey} found in Modflow 6 model, "
-                "but Sprinkling package found in MetaSWAP. "
-                "iMOD Coupler requires a Well Package "
-                "to couple wells."
-            )
-        elif not sprinkling_in_msw and sprinkling_in_mf6:
-            raise ValueError(
-                f"Modflow 6 Well package {mf6_wel_pkgkey} specified for sprinkling, "
-                "but no Sprinkling package found in MetaSWAP model."
-            )
-        elif sprinkling_in_msw and sprinkling_in_mf6:
-            return True
+            svat_id = svat_data[:, 1]
+            svat_lay = svat_data[:, 2]
+            for vi in range(svat_id.size):
+                svat_lookup[(svat_id[vi], svat_lay[vi])] = vi
         else:
-            return False
+            raise ValueError(f"Can't find {msw_mod2svat_file}.")
 
-    def write(
-        self,
-        directory: Union[str, Path],
-        modflow6_dll: Union[str, Path],
-        metaswap_dll: Union[str, Path],
-        metaswap_dll_dependency: Union[str, Path],
-        dflowfm_dll: Union[str, Path],
-    ):
-        """
-        Write MetaSWAP and Modflow 6 model with exchange files, as well as a
-        ``.toml`` file which configures the imod coupler run.
-
-        Parameters
-        ----------
-        directory: str or Path
-            Directory in which to write the coupled models
-        modflow6_dll: str or Path
-            Path to modflow6 .dll. You can obtain this library by downloading
-            `the last iMOD5 release
-            <https://oss.deltares.nl/web/imod/download-imod5>`_
-        metaswap_dll: str or Path
-            Path to metaswap .dll. You can obtain this library by downloading
-            `the last iMOD5 release
-            <https://oss.deltares.nl/web/imod/download-imod5>`_
-        metaswap_dll_dependency: str or Path
-            Directory with metaswap .dll dependencies. Directory should contain:
-            [fmpich2.dll, mpich2mpi.dll, mpich2nemesis.dll, TRANSOL.dll]. You
-            can obtain these by downloading `the last iMOD5 release
-            <https://oss.deltares.nl/web/imod/download-imod5>`_
-        dflowfm_dll: str or Path
-            Directory with dflowfm.dll as well as all of its dependencies. 
-        """
-        # force to Path
-        directory = Path(directory)
-        # For some reason the Modflow 6 model has to be written before
-        # writing the MetaSWAP model. Else we get an Access Violation Error when
-        # running the coupler.
-        self.dfm_model.save(directory / self._dfm_model_dir/ "dfm.mdu", recurse=True)
-        self.mf6_simulation.write(directory / self._modflow6_model_dir)
-        self.msw_model.write(directory / self._metaswap_model_dir)
-
-
-        # Write exchange files
-        exchange_dir = directory / "exchanges"
-        exchange_dir.mkdir(mode=755, exist_ok=True)
-        self.write_exchanges(exchange_dir, self.mf6_rch_pkgkey, self.mf6_wel_pkgkey)
-
-        coupling_dict = self._get_coupling_dict(
-            exchange_dir, self.mf6_rch_pkgkey, self.mf6_wel_pkgkey
+        # create mappings
+        table_node2svat: NDArray[np.int32] = np.loadtxt(
+            self.coupling.mf6_msw_node_map, dtype=np.int32, ndmin=2
         )
-
-        self.write_toml(
-            directory,
-            modflow6_dll,
-            metaswap_dll,
-            metaswap_dll_dependency,
-            dflowfm_dll,
-            coupling_dict,
-        )
-
-    def write_toml(
-        self,
-        directory: Union[str, Path],
-        modflow6_dll: Union[str, Path],
-        metaswap_dll: Union[str, Path],
-        metaswap_dll_dependency: Union[str, Path],
-        dflowfm_dll: Union[str, Path],       
-        coupling_dict: dict,
-    ):
-        """
-        Write .toml file which configures the imod coupler run.
-
-        Parameters
-        ----------
-        directory: str or Path
-            Directory in which to write the .toml file.
-        modflow6_dll: str or Path
-            Path to modflow6 .dll. You can obtain this library by downloading
-            `the last iMOD5 release
-            <https://oss.deltares.nl/web/imod/download-imod5>`_
-        metaswap_dll: str or Path
-            Path to metaswap .dll. You can obtain this library by downloading
-            `the last iMOD5 release
-            <https://oss.deltares.nl/web/imod/download-imod5>`_
-        metaswap_dll_dependency: str or Path
-            Directory with metaswap .dll dependencies. Directory should contain:
-            [fmpich2.dll, mpich2mpi.dll, mpich2nemesis.dll, TRANSOL.dll]. You
-            can obtain these by downloading `the last iMOD5 release
-            <https://oss.deltares.nl/web/imod/download-imod5>`_
-        coupling_dict: dict
-            Dictionary with names of coupler packages and paths to mappings.
-        """
-        # force to Path
-        directory = Path(directory)
-
-        toml_path = directory / self._toml_name
-
-        coupler_toml = {
-            "timing": False,
-            "log_level": "INFO",
-            "driver_type": "dfm_metamod",
-            "driver": {
-                "kernels": {
-                    "modflow6": {
-                        "dll": str(modflow6_dll),
-                        "work_dir": f".\\{self._modflow6_model_dir}",
-                    },
-                    "metaswap": {
-                        "dll": str(metaswap_dll),
-                        "work_dir": f".\\{self._metaswap_model_dir}",
-                        "dll_dep_dir": str(metaswap_dll_dependency),
-                    },
-                    "dflowfm": {
-                        "dll": str(dflowfm_dll),
-                        "work_dir": f".\\{self._dfm_model_dir}",
-                    },                    
-                },
-                "coupling": [coupling_dict],
-            },
-        }
-
-        with open(toml_path, "wb") as f:
-            tomli_w.dump(coupler_toml, f)
-
-    def _get_gwf_modelnames(self):
-        """
-        Get names of gwf models in mf6 simulation
-        """
-        return [
-            key
-            for key, value in self.mf6_simulation.items()
-            if value._pkg_id == "model"
+        node_idx = table_node2svat[:, 0] - 1
+        msw_idx = [
+            svat_lookup[table_node2svat[ii, 1], table_node2svat[ii, 2]]
+            for ii in range(len(table_node2svat))
         ]
 
-    def _get_dfm_modelname(self):    
-        return "dfm.mdu"    
+        self.map_msw2mod["storage"], self.mask_msw2mod["storage"] = create_mapping(
+            msw_idx,
+            node_idx,
+            self.msw_storage.size,
+            self.mf6_storage.size,
+            "sum",
+        )
 
-    def _get_coupling_dict(
-        self,
-        directory: Union[str, Path],
-        mf6_rch_pkgkey: str,
-        mf6_wel_pkgkey: Optional[str],
-    ) -> dict:
-        """
-        Get dictionary with names of coupler packages and paths to mappings.
+        # MetaSWAP gives SC1*area, MODFLOW by default needs SS, convert here.
+        # When MODFLOW is configured to use SC1 explicitly via the
+        # STORAGECOEFFICIENT option in the STO package, only the multiplication
+        # by area needs to be undone
+        if self.mf6_has_sc1:
+            conversion_terms = 1.0 / self.mf6_area
+        else:
+            conversion_terms = 1.0 / (self.mf6_area * (self.mf6_top - self.mf6_bot))
 
-        Parameters
-        ----------
-        directory: str or Path
-            Directory where .dxc files are written.
-        mf6_rch_pkgkey: str
-            Key of Modflow 6 recharge package to which MetaSWAP is coupled.
-        mf6_wel_pkgkey: str
-            Key of Modflow 6 well package to which MetaSWAP sprinkling is
-            coupled.
+        conversion_matrix = dia_matrix(
+            (conversion_terms, [0]),
+            shape=(self.mf6_area.size, self.mf6_area.size),
+            dtype=self.mf6_area.dtype,
+        )
+        self.map_msw2mod["storage"] = conversion_matrix * self.map_msw2mod["storage"]
 
-        Returns
-        -------
-        coupling_dict: dict
-            Dictionary with names of coupler packages and paths to mappings.
-        """
+        self.map_mod2msw["head"], self.mask_mod2msw["head"] = create_mapping(
+            node_idx,
+            msw_idx,
+            self.mf6_head.size,
+            self.msw_head.size,
+            "avg",
+        )
 
-        coupling_dict = {}
+        table_rch2svat: NDArray[np.int32] = np.loadtxt(
+            self.coupling.mf6_msw_recharge_map, dtype=np.int32, ndmin=2
+        )
+        rch_idx = table_rch2svat[:, 0] - 1
+        msw_idx = [
+            svat_lookup[table_rch2svat[ii, 1], table_rch2svat[ii, 2]]
+            for ii in range(len(table_rch2svat))
+        ]
 
-        gwf_names = self._get_gwf_modelnames()
+        self.map_msw2mod["recharge"], self.mask_msw2mod["recharge"] = create_mapping(
+            msw_idx,
+            rch_idx,
+            self.msw_volume.size,
+            self.mf6_recharge.size,
+            "sum",
+        )
 
-        coupling_dict["dfm_model"] = self._get_dfm_modelname()
+        if self.coupling.enable_sprinkling:
+            assert isinstance(self.coupling.mf6_msw_well_pkg, str)
+            assert isinstance(self.coupling.mf6_msw_sprinkling_map, Path)
 
-        # Assume only one groundwater flow model
-        # FUTURE: Support multiple groundwater flow models.
-        coupling_dict["mf6_model"] = gwf_names[0]
+            # in this case we have a sprinkling demand from MetaSWAP
+            mf6_sprinkling_tag = self.mf6.get_var_address(
+                "BOUND", self.coupling.mf6_model, self.coupling.mf6_msw_well_pkg
+            )
+            self.mf6_sprinkling_wells = self.mf6.get_value_ptr(mf6_sprinkling_tag)[:, 0]
+            table_well2svat: NDArray[np.int32] = np.loadtxt(
+                self.coupling.mf6_msw_sprinkling_map, dtype=np.int32, ndmin=2
+            )
+            well_idx = table_well2svat[:, 0] - 1
+            msw_idx = [
+                svat_lookup[table_well2svat[ii, 1], table_well2svat[ii, 2]]
+                for ii in range(len(table_well2svat))
+            ]
 
-        coupling_dict[
-            "mf6_msw_node_map"
-        ] = f"./{directory.name}/{NodeSvatMapping._file_name}"
+            (
+                self.map_msw2mod["sprinkling"],
+                self.mask_msw2mod["sprinkling"],
+            ) = create_mapping(
+                msw_idx,
+                well_idx,
+                self.msw_volume.size,
+                self.mf6_sprinkling_wells.size,
+                "sum",
+            )
 
-        coupling_dict["mf6_msw_recharge_pkg"] = mf6_rch_pkgkey
-        coupling_dict[
-            "mf6_msw_recharge_map"
-        ] = f"./{directory.name}/{RechargeSvatMapping._file_name}"
+    def update(self) -> None:
+        # heads to MetaSWAP
+        self.exchange_mod2msw()
 
-        coupling_dict["enable_sprinkling"] = self.is_sprinkling
+        # we cannot set the timestep (yet) in Modflow
+        # -> set to the (dummy) value 0.0 for now
+        self.mf6.prepare_time_step(0.0)
 
-        if self.is_sprinkling:
-            coupling_dict["mf6_msw_well_pkg"] = mf6_wel_pkgkey
-            coupling_dict[
-                "mf6_msw_sprinkling_map"
-            ] = f"./{directory.name}/{WellSvatMapping._file_name}"
+        self.delt = self.mf6.get_time_step()
+        self.msw.prepare_time_step(self.delt)
 
-        return coupling_dict
+        # convergence loop
+        self.mf6.prepare_solve(1)
+        for kiter in range(1, self.max_iter + 1):
+            has_converged = self.do_iter(1)
+            if has_converged:
+                logger.debug(f"MF6-MSW converged in {kiter} iterations")
+                break
+        self.mf6.finalize_solve(1)
 
-    def write_exchanges(
-        self,
-        directory: Union[str, Path],
-        mf6_rch_pkgkey: str,
-        mf6_wel_pkgkey: Optional[str],
-    ):
-        """
-        Write exchange files (.dxc) which map MetaSWAP's svats to Modflow 6 node
-        numbers, recharge ids, and well ids.
+        self.mf6.finalize_time_step()
+        self.msw_time = self.mf6.get_current_time()
+        self.msw.finalize_time_step()
 
-        Parameters
-        ----------
-        directory: str or Path
-            Directory where .dxc files are written.
-        mf6_rch_pkgkey: str
-            Key of Modflow 6 recharge package to which MetaSWAP is coupled.
-        mf6_wel_pkgkey: str
-            Key of Modflow 6 well package to which MetaSWAP sprinkling is
-            coupled.
-        """
+    def finalize(self) -> None:
+        self.mf6.finalize()
+        self.msw.finalize()
 
-        gwf_names = self._get_gwf_modelnames()
+    def get_current_time(self) -> float:
+        return self.mf6.get_current_time()
 
-        # Assume only one groundwater flow model
-        # FUTURE: Support multiple groundwater flow models.
-        gwf_model = self.mf6_simulation[gwf_names[0]]
+    def get_end_time(self) -> float:
+        return self.mf6.get_end_time()
 
-        grid_data_key = [
-            pkgname
-            for pkgname, pkg in self.msw_model.items()
-            if isinstance(pkg, GridData)
-        ][0]
+    def exchange_msw2mod(self) -> None:
+        """Exchange Metaswap to Modflow"""
+        self.mf6_storage[:] = (
+            self.mask_msw2mod["storage"][:] * self.mf6_storage[:]
+            + self.map_msw2mod["storage"].dot(self.msw_storage)[:]
+        )
 
-        dis = gwf_model[gwf_model._get_pkgkey("dis")]
+        # Divide recharge and extraction by delta time
+        tled = 1 / self.delt
+        self.mf6_recharge[:] = (
+            self.mask_msw2mod["recharge"][:] * self.mf6_recharge[:]
+            + tled * self.map_msw2mod["recharge"].dot(self.msw_volume)[:]
+        )
 
-        index, svat = self.msw_model[grid_data_key].generate_index_array()
-        grid_mapping = NodeSvatMapping(svat, dis)
-        grid_mapping.write(directory, index, svat)
+        if self.coupling.enable_sprinkling:
+            self.mf6_sprinkling_wells[:] = (
+                self.mask_msw2mod["sprinkling"][:] * self.mf6_sprinkling_wells[:]
+                + tled * self.map_msw2mod["sprinkling"].dot(self.msw_volume)[:]
+            )
 
-        recharge = gwf_model[mf6_rch_pkgkey]
+    def exchange_mod2msw(self) -> None:
+        """Exchange Modflow to Metaswap"""
+        self.msw_head[:] = (
+            self.mask_mod2msw["head"][:] * self.msw_head[:]
+            + self.map_mod2msw["head"].dot(self.mf6_head)[:]
+        )
 
-        rch_mapping = RechargeSvatMapping(svat, recharge)
-        rch_mapping.write(directory, index, svat)
+    def do_iter(self, sol_id: int) -> bool:
+        """Execute a single iteration"""
+        self.msw.prepare_solve(0)
+        self.msw.solve(0)
+        self.exchange_msw2mod()
+        has_converged = self.mf6.solve(sol_id)
+        self.exchange_mod2msw()
+        self.msw.finalize_solve(0)
+        return has_converged
 
-        if self.is_sprinkling:
-            well = gwf_model[mf6_wel_pkgkey]
-            well_mapping = WellSvatMapping(svat, well)
-            well_mapping.write(directory, index, svat)
+    def report_timing_totals(self) -> None:
+        total_mf6 = self.mf6.report_timing_totals()
+        total_msw = self.msw.report_timing_totals()
+        total = total_mf6 + total_msw
+        logger.info(f"Total elapsed time in numerical kernels: {total:0.4f} seconds")
