+++ tests/test_dfm_metamod_cases.py
@@ -1,5 +1,5 @@
 import pytest
-from imod_coupler.drivers.dfm_metamod.dfm_metamod import Dfm_MetaMod
+from imod_coupler.drivers.dfm_metamod.dfm_metamod_model import DfmMetaModModel
 from imod.mf6 import Modflow6Simulation
 from imod.msw import MetaSwapModel
 from hydrolib.core.io.mdu.models import FMModel
@@ -12,9 +12,9 @@ def case_default(
     coupled_mf6_model: Modflow6Simulation,
     prepared_msw_model: MetaSwapModel,
     prepared_dflowfm_model: FMModel,
-)-> Dfm_MetaMod:
+)-> DfmMetaModModel:
 
-    default = Dfm_MetaMod(
+    default = DfmMetaModModel(
         prepared_msw_model,
         coupled_mf6_model,
         prepared_dflowfm_model,
